Move LexicalStates in jflex.scanner
