+++ jflex/src/main/java/jflex/core/BUILD
@@ -14,5 +14,6 @@ java_library(
         "//jflex/src/main/java/jflex/io",
         "//jflex/src/main/java/jflex/l10n",
         "//jflex/src/main/java/jflex/performance",
+        "//jflex/src/main/java/jflex/scanner",
     ],
 )
