+++ jflex/src/main/java/jflex/scanner/BUILD
@@ -1,0 +1,6 @@
+package(default_visibility = ["//jflex:visibility"])
+
+java_library(
+    name = "scanner",
+    srcs = glob(["*.java"]),
+)
