+++ jflex/src/main/java/jflex/scanner/LexicalStates.java
@@ -7,7 +7,7 @@
  *                                                                         *
  * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * */
 
-package jflex.core;
+package jflex.scanner;
 
 import java.util.ArrayList;
 import java.util.LinkedHashMap;
